Rename DOOM app directories
