+++ apps/doom-mcp-server/package.json
