+++ apps/doom-mcp-server/src/config.ts
