+++ apps/doom-mcp-server/src/mcp/schemas.ts
