+++ apps/doom-mcp-server/src/mcp/server.ts
