+++ apps/doom-mcp-server/src/mcp/tools.ts
