+++ apps/doom-mcp-server/src/mcp/validator.ts
