+++ apps/doom-mcp-server/src/mcp/zodSchemas.ts
