+++ apps/doom-mcp-server/src/protocol/jsonRpc.ts
