+++ apps/doom-mcp-server/src/server.ts
