+++ apps/doom-mcp-server/src/session/bootstrap.ts
