+++ apps/doom-mcp-server/src/tests/token.test.ts
