+++ apps/doom-mcp-server/src/tests/tools.test.ts
