+++ doom/scripts/doom-netlify-sync.mjs
@@ -3,7 +3,7 @@ import { join } from 'node:path'
 import { fileURLToPath } from 'node:url'
 
 const rootDir = fileURLToPath(new URL('../..', import.meta.url))
-const doomDistDir = join(rootDir, 'apps', 'web', 'dist')
+const doomDistDir = join(rootDir, 'apps', 'doom-web', 'dist')
 const publicDir = join(rootDir, 'public')
 
 const targets = [join(publicDir, 'doom')]
