+++ doom/tsconfig.json
@@ -11,8 +11,8 @@
     "noEmitOnError": true
   },
   "include": [
-    "../apps/web/src/**/*.ts",
-    "../apps/mcp-server/src/**/*.ts",
+    "../apps/doom-web/src/**/*.ts",
+    "../apps/doom-mcp-server/src/**/*.ts",
     "../packages/doom-adapter/src/**/*.ts",
     "../packages/doom-session/src/**/*.ts",
     "../netlify/functions/doom-*.mts",
